+++ CHANGELOG.md
@@ -30,3 +30,5 @@ $ /usr/bin/env php bin/console darvin:utils:translations:title-case <pathname>
 ```
 
 6.4.6: Remove "symfony/symfony" from dependencies.
+
+6.5.0: Add "format_price" Twig filter.